requirements.txt
langchain==0.3.0
langchain-core>=0.3.33,<0.4.0
langchain-community==0.3.0
openai>=1.58.1,<2.0.0
tiktoken
python-dotenv
pydantic==2.9.2
pydantic-core==2.23.4
langchain-deepseek
numpy==1.26.4
requests==2.32.3
